More support for LLVM-3.6.
In particular:

In Trace, there is code for identifying the line of source code
corresponding to a certain LLVM IR instruction. This code contains a
fallback which is only used when the LLVM IR input file contains
malformed metadata (caused by the compiler). This fallback is now
disabled in case it is incompatible with the LLVM headers.

Metadata is no longer considered to be values. This is mirrored in
changes in the LLVM assembly syntax. Test cases which are affected
have been rewritten.