Bias --dump-spec towards earlier edges over later
This is an attempted fix aimed at a perceived flaw in --dump-spec, in
that it generates specs that are "technically correct", by only matching
one trace due to containing IIDs that are unique to that trace, but
"miss the point" in what characterises the trace.

This change does sometimes result in larger specs, but hopefully closer
to what a human would write.