Use memcpy for nondeterminism tests
Optimal-DPOR needs to be able to reason about which events that will
conflict in future executions, and therefore disallows programs that
pull addresses out of thin air. Therefore, use memcpy() to mask the
accesses to addresses that do not belong to the program so that the test
program is not rejected too early.