Use pkg-config to autodetect flags for libffi
By default libffi is installed outside the include path, and multiple
Linux distributions (Arch & Gentoo, among others) do not override this.
Instead of trying to hard-code all possible locations it could be
installed in configure, use pkg-config to get the right include
directory directly.

The old library detection has been left as a fallback, should there be
systems without pkg-config.

This might subsume "AC_CHECK_HEADERS([ffi/ffi.h], ..." later in the
script, but since I don't know which system this was introduced to
support, I cannot test if this is the case, and so I have left it be.
The order has been permuted so that the standard "#include <ffi.h>" is
preferred, however.