Remove AC_FUNC_MALLOC from configure
AC_FUNC_MALLOC is harmful. First of all, it is fundamentally
incompatible with C++ (since it does "#define malloc rpl_malloc" when it
takes issue with the system malloc or something else in the
configuration). Second of all, even when used for C, a project using it
must contain a "malloc.c" that provides a fallback implementation, which
nidhugg is not.

In practice, it only serves to mask misconfiguration errors by causing
strange build errors in the c++ standard library ("error: ‘::malloc’ has
not been declared"), so remove it.