Fix ambiguity in nidhuggc command line parsing
5db3df3 introduced an ambiguity in the nidhuggc command line syntax when
passing compiler or nidhugg options that take arguments after a
space (for example "-I ." rather than "-I.").

Additionally, the parsing was broken in the sense that options that are
understood by nidhuggc, such as "--help", could not be passed as program
arguments.

The ambiguity arises because both the "--" separated sections of
compiler and program arguments were made optional, with every
combination accepted. In the case of a single "--", to disambiguate
between having compiler&nidhugg args or having nidhugg&program args,
nidhuggc looked at if the first section contained an argument that
looked like an input file, and if so treated it as nidhugg arguments,
and otherwise as compiler arguments. However, this only works based on
the assumption that all compiler arguments start with a "-"

This ambiguity is resolved by disallowing the "nidhuggc file.c --
[PROGRAM ARGS]" short form.

Additionally, we update the synopsis line in the help message to be the
same as the synopsis in the manual, which only describes a subset of
allowed forms but makes it clear that the removed short form is
disallowed.

Fixes #28.