SmtlibSatSolver: Add flag for encoding with bitvectors
A new flag, --smtlib-bv, switches the encoding to use bitvectors. This
allows the use of bitvector-only SMT solvers such as boolector and STP.

STP does not yet work, though, as it does not support
"(get-value (...))", only "(get-model)", and also cannot parse the
"(set-option :produce-models true)" directive, which is required by some
other solvers.