SmtlibSatSolver: Add flag to ask for QF_LIA logic
Yices does not like that we ask for the QF_IDL logic, since per the
specification it is not allowed to contain (or ...). However, just
reporting QF_LIA in all cases is not desirable, as at least Z3 becomes
much slower with this logic, even if the formula is identical.