Fix global initialization order fiasco with Timing
The symptom of this bug was that some contexts were missing from the
timing report.

You cannot rely on the order of construction of global variables;
therefore you cannot assume that a std::vector global is initialised in
a constructor also called as part of global initialisation. Therefore,
replace the std::vector with a manual linked list to fix the bug.