TSOTraceBuilder: Fix merging different sizes in WuT
When inserting into the wakeup tree, the code would not consider the
sizes of the events. If a sequence where a big event had been split into
several smaller ones was merged with an existing branch where the big
event existed, only the first of the split events would be removed from
the wakeup sequence, and the latter ones would incorrectly remain and
cause nonexistent program steps to be replayed.