Fix edge case in nondet detection for PSO
If the initial event from the backtrack set was a branch (control flow),
the nondeterminism detection from ac12d1ab would mistakenly assume that
it must go the same way (true/false) as the next branch in the previous
execution of the program, which might be by a different thread entirely.
This can cause spurious nondeterminism error reports.

Already fixed for SC&TSO in 702d4388. Does not affect POWER/ARM.