Introduce new wrapper type AnyCallInst
LLVM starting with version 11 does not have the llvm::CallSite helper
that previously served this purpose, which we use in AnyCallInst on
older LLVMs that don't have llvm::CallBase. The interface is similar-ish
to that offered by a (llvm::CallBase&), except that it is nullable,
which should make migration easier once we drop those old llvm versions.