Partial Loop Purity Elimination pass
This pass generalises and subsumes the SpinAssume pass.

Partial Loop Purity Elimination discovers paths through loops that are
"pure"; that do not have observable side effects and returns to the same
local state as before the loop iteration. Such pure loop paths are then
eliminated by inserting assume-statements.

It works by for each loop computing a set of binary predicates over LLVM
values such that if any of them hold at any point in the loop, then the
current loop iteration is pure. This set is computed through a backwards
dataflow analysis.

Loops that contain function calls are analysed by selectively inlining
calls in loops that might be pure prior to the analysis.