Consistently use braces around `else`
Enforce the rule that an `else` should either have braces on both its
sides or none of them.

While at it, fix some whitespace issues and add a missing include.

There is one more place in the code that the above rule is violated
(src/PartialLoopPurityPass.cpp:325), which is left for @margnus1 to fix.