configure.ac: fix finding dlopen() on OpenBSD
AC_CHECK_LIB checks for the presence of a given function in a library,
and fails if the library can't be found or doesn't contain the function..

AC_SEARCH_LIBS checks if a function is present, and if it isn't, tries
again with each specified library in sequence, stopping and linking in
the first library to contain the function, and failing if the function
isn't found anywhere.

Linux systems have dlopen inside a libdl library. OpenBSD has dlopen
inside libc, this change fixes compilation on OpenBSD and probably the
other BSDs.

Ref: https://www.gnu.org/software/autoconf/manual/autoconf-2.71/html_node/Libraries.html