Change mentions of SWSC to Nidhugg
The name SWSC is a relik that is confusing to both developers and users.
Also, it's a bit strange to have copyright messages that claim that
"This test/benchmark is part of SWSC" in Nidhugg's code base.

This commit changes these to mention Nidhugg instead.

While at it, performed some minor cleanups in some tests and benchmarks.