Make `LLVMUtils::getPthreadTType` smarter
If we have definitions of `pthread_join` or `pthread_self`
look them up and take the definition from there.

This is needed when the IR module comes from another system where the
pthread type is different. Or, as in what motivated this change, we have
hardcoded IR in unit tests and want those tests to work everywhere.

Thanks to @margnus1 for this commit.