Provide an alternative to `ConstantExpr::getPredicate`
LLVM 19 removed the `getPredicate` method from `ConstExpr`; in fact,
it's quite possible that a whole bunch of ConstExpr instructions were
removed or are not generated anymore as e.g. can be seen by comparing
the current [Execution.cpp](https://github.com/llvm/llvm-project/blob/30c49a40222a8e2f90565370ab20253c1e426383/llvm/lib/ExecutionEngine/Interpreter/Execution.cpp)
file with the one in Nidhugg's code base.

The current commit/change makes an attempt to provide an alternative way
to extract the predicate of an `ICmp` and a `FCmp` instruction used as a
constant expression if we want to keep that code -- although it is not
triggered by any test in Nidhugg's test suite.  An alternative fix would
be to if-def out or even completely remove the `FCmp` and `ICmp` cases
from the switch.

The PR is submitted separate here in order to discuss which alternative
we prefer.